fix login expert and fix notify

+++ User-Service/src/main/java/com/se1/userservice/domain/payload/UserResponseDto.java
@@ -23,7 +23,7 @@ public class UserResponseDto {
     private Boolean isExpert;
     private String topicId;
 	private Double rating;
-	private Long ratingCount;
+	private Integer ratingCount;
 	@JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
     private Date lastTime;
 }
